+++ src/layout/List/ListTheme.ts
@@ -53,13 +53,13 @@ const baseTheme: ListTheme = {
 
 export const listTheme = {
   ...baseTheme,
-  base: [baseTheme.base, 'text-surface-content'].join(' '),
-  header: [baseTheme.header, 'text-surface-content'].join(' '),
+  base: [baseTheme.base, 'text-text-primary'].join(' '),
+  header: [baseTheme.header, 'text-text-primary'].join(' '),
   listItem: {
     ...baseTheme.listItem,
     base: [
       baseTheme.listItem.base,
-      'hover:bg-panel-accent hover:text-mystic light:hover:bg-vulcan/5 light:hover:text-panel-secondary-content [&:has(h3)]:hover:bg-transparent'
+      'hover:bg-panel-accent hover:text-mystic light:hover:bg-vulcan/5 light:hover:text-text-secondary [&:has(h3)]:hover:bg-transparent'
     ].join(' '),
     active: [
       baseTheme.listItem.active,
@@ -67,7 +67,7 @@ export const listTheme = {
     ].join(' '),
     disabled: [
       baseTheme.listItem.disabled,
-      'opacity-40 text-panel-secondary-content'
+      'opacity-40 text-text-secondary'
     ].join(' ')
   }
 };